+++ src/lib/util.ts
@@ -1,11 +1,45 @@
 import * as vscode from 'vscode'
 
-export function get_config(key: string): any {
-	return vscode.workspace.getConfiguration('renpyWarp').get(key)
+import { name as extension_name } from '../../package.json'
+
+export function get_configuration_object(): vscode.WorkspaceConfiguration {
+	return vscode.workspace.getConfiguration(extension_name)
 }
 
-export function set_config(key: string, value: any): void {
-	vscode.workspace
-		.getConfiguration('renpyWarp')
-		.update(key, value, vscode.ConfigurationTarget.Global)
+export function get_config(key: string): any {
+	return vscode.workspace.getConfiguration(extension_name).get(key)
+}
+
+export async function set_config(
+	key: string,
+	value: any,
+	workspace = false
+): Promise<void> {
+	return vscode.workspace
+		.getConfiguration(extension_name)
+		.update(
+			key,
+			value,
+			workspace
+				? vscode.ConfigurationTarget.Workspace
+				: vscode.ConfigurationTarget.Global
+		)
+}
+
+/**
+ * sets a configuration value exclusively for the workspace or globally,
+ * removing the other setting if it exists
+ */
+export async function set_config_exclusive(
+	key: string,
+	value: any,
+	workspace = false
+): Promise<void> {
+	await set_config(key, value, workspace)
+	await set_config(key, undefined, !workspace)
+}
+
+export async function show_file(path: string): Promise<void> {
+	const doc = await vscode.workspace.openTextDocument(path)
+	await vscode.window.showTextDocument(doc)
 }